+++ anstoss/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class AnstossConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "anstoss"
+    verbose_name = "Anstöße (Feedback)"